% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials.R
\docType{data}
\name{trials}
\alias{trials}
\title{IACGMOOH bushtucker trials data}
\format{
\subsection{\code{trials}}{

A data frame with 506 rows and 8 columns:
\describe{
\item{season}{(\emph{numeric}) Season of the show}
\item{number}{(\emph{numeric}) Number of the trial (i.e. 1 = the first trial of the season,
2 = second trial of the season, etc.)}
\item{date}{(\emph{Date}) Date of the trial}
\item{name}{(\emph{character}) Name of the trial}
\item{live}{(\emph{logical}) TRUE if the trial was broadcast live, FALSE otherwise}
\item{chosen_by}{(\emph{character}) Who chose the contestants for the trial? One of:
'public' if it was by public/viewer vote; 'contestants' if the campmates chose amongst themselves;
or 'showrunners' if the trial was compulsory for the chosen contestants}
\item{contestant}{(\emph{character}) First name(s) of the contestant(s) that took part in the trial.
Where there are multiple contestants, they are separated by a semi-colon.}
\item{stars_available}{(\emph{numeric}) Maximum number of stars available to win in this trial}
\item{stars_won}{(\emph{numeric}) Number of stars that the contestant(s) won in the trial}
}
}
}
\source{
\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_1}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_2}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_3}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_4}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_5}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_6}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_7}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_8}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_9}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_10}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_11}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_12}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_13}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_14}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_15}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_16}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_17}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_18}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_19}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_20}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_21}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_22}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_23}

\url{https://en.wikipedia.org/wiki/I\%27m_a_Celebrity...Get_Me_Out_of_Here!_(British_TV_series)_series_24}
}
\usage{
trials
}
\description{
Data on each of the viewer votes in the first 24 seasons of the
UK TV show "I'm A Celebrity, Get Me Out Of Here"
}
\keyword{datasets}
